+++ src/ic-cdk/src/api/management_canister/mod.rs
@@ -14,4 +14,4 @@ pub mod bitcoin;
 pub mod http_request;
 pub mod main;
 pub mod provisional;
-pub mod threshold_ecdsa;
+pub mod ecdsa;
